+++ plugin/src/main/groovy/org/openbakery/UIAutomationTestExtension.groovy
@@ -1,0 +1,12 @@
+package org.openbakery
+
+/**
+ *
+ * @author René Pirringer
+ *
+ */
+class UIAutomationTestExtension {
+
+	String scriptsDirectory = "test/uiautomation"
+
+}