+++ plugin/src/main/groovy/org/openbakery/XcodeBuildCleanTask.groovy
@@ -16,6 +16,7 @@
 package org.openbakery
 
 import org.gradle.api.DefaultTask
+import org.gradle.api.internal.ConventionTask
 import org.gradle.api.tasks.TaskAction